azure-certifications-study-guide/AI-projects/project-05-visual-search/src/requirements.txt
azure-functions==1.17.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
